requirements.txt
streamlit
Pillow
pandas
sklearn
joblib
numpy
matplotlib